+++ src/lib.rs
@@ -263,6 +263,11 @@ pub enum Error {
     BadEoc,
     /// The contents include invalid octet(s).
     InvalidContents,
+    /// IO Error for serialization/deserialization.
+    ///
+    /// Note that this error cannot be compared with others.
+    /// `PartialEq::eq` always returns `false` for this error.
+    Io(std::io::Error),
 }
 
 impl fmt::Display for Error {
@@ -274,6 +279,7 @@ impl fmt::Display for Error {
             Self::IndefiniteLength => f.write_str("'Indefinite Length' in DER or CER"),
             Self::BadEoc => f.write_str("'Indefinite Length BER' includes a bad 'EOC.'"),
             Self::InvalidContents => f.write_str("Contents include invlid octet(s)."),
+            Self::Io(err) => err.fmt(f),
         }
     }
 }
@@ -289,6 +295,7 @@ impl PartialEq for Error {
             Self::IndefiniteLength => matches!(other, Self::IndefiniteLength),
             Self::BadEoc => matches!(other, Self::BadEoc),
             Self::InvalidContents => matches!(other, Self::InvalidContents),
+            Self::Io(_) => false,
         }
     }
 }